+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import App from "./Router";
+import Router from "./Router";
 import { IssuesProvider } from "./context/issuesContext";
 import { IssueDetailProvider } from "./context/issueDetailContext";
 
@@ -8,7 +8,7 @@ const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <IssuesProvider>
     <IssueDetailProvider>
-      <App />
+      <Router />
     </IssueDetailProvider>
   </IssuesProvider>,
 );
