+++ src/pages/IssueDetail.js
@@ -1,0 +1,12 @@
+import Layout from "../components/common/Layout";
+import IssueDetailItem from "../components/issue/IssueDetailItem/IssueDetailItem";
+
+const IssueDetail = () => {
+  return ( 
+    <Layout>
+      <IssueDetailItem />
+    </Layout>
+   );
+}
+ 
+export default IssueDetail;
