--- src/pages/issueDetail/IssueDetail.js
@@ -1,12 +1,0 @@
-import Layout from "../../components/common/Layout";
-import IssueDetailItem from "../../components/issue/IssueDetailItem/IssueDetailItem";
-
-const IssueDetail = () => {
-  return ( 
-    <Layout>
-      <IssueDetailItem />
-    </Layout>
-   );
-}
- 
-export default IssueDetail;
