--- src/pages/issueList/Style.js
@@ -1,7 +1,0 @@
-import styled from "@emotion/styled";
-
-export const IssueListContainer = styled.section`
-  width: 70vw;
-  height: 100%;
-  margin: 0 auto;
-`;
