--- src/style/globalStyle.js
@@ -1,3 +1,0 @@
-import { css } from "@emotion/react";
-
-export const customBody = css``;
